Added possibility to register comparators to the type system.
The reason for that is following:
Template types might define the == operator, but not the contained template type.
e.g std::vector<foo>
With SFINAE, it is not possible to check if the contained template paramter itself has the comparison operator defined.
Or other way spoken, whether the call to the comparision operator is valid or make a compile error.

It is also unclear, whether the contained template parameter is needed for comparison or not.
We can not provide a 100% solution for this problem here.
So we make it explicit, the user has to provide the comparator operator for such template types. (EXCEPT: std::string)

- added unit test
- added documentation