Added new class "array_range" and "flat_map"
The class "array_range" is used to get a view into the stored properties,
instead copying all properties together into a std::vector, which allocates heap memory.

Remark: The memory usage will be slightly higher, because all properties including base properties
of a type are now stored inside an own vector.

The class "flat_map" is used for internal usage; it is a map with contiguous memory for key and values.

This branch is to implement the request of issue #6

To Do: Implement the same for methods, constructors, enumerations and types.