Adjusted return value of get_parameter_infos() function.
The method and constructor class will now return a range of paramter_info objects,
instead of a vector.

Therefore an additional std::array<parameter_info> object needs to be stored inside the method object.
However, because parameter_info is no default constructable, it is not possible to create a std::array<parameter_info, 0>
This is an issue in the c++ standard right now, see the discussion here: https://cplusplus.github.io/LWG/lwg-active.html#2157