fixed invalid access of memory, reported by valgrind.
The reason for the invalid access reported by valgrind, was because of an already removed
std::string in "custom_names" which was accessed via the string_view in "m_custom_name_to_id".
There were two string_views with the same name inserted into the flat_map. Which should not be allowed.

Therefore the flat_map class is now really a map, no double entires. And a separate class flat_multimap was created.

To really fix the memory access error, I switched to use std::string now for the mapping of custom name to id.
Which increases the memory usage, but makes the memory management easier.