ATTENTION!: Replaced default constructor binding from "as_object" to "as_std_shared_ptr".
This might break some code silently!
Without this change it is not possible to register a constructor
from a class which has a deleted or private copy constructor.

This change will allow now the registration of these kind of classes.

shared_ptr is a valid choice because it has automatic memory management.
Its better here to be explicit and have one rule, instead of doing some hidden magic for this case.

To have the old binding behaviour back, add the policy: 'policy::ctor::as_object'
to every constructor
registration::class_<MyStruct>("MyStruct")
    .constructor<>()
    (
        policy::ctor::as_object
    );

Remark:
When you want to real type of the created instance in the variant, use 'type::get_wrapped_type()'

This will fix issue #14