Fixed possible dereferencing nullptr in array_range<T>
When the range is empty, calling begin() could lead to a possible crash,
which should not happen. The problem was, the value was dereferenced and forwarded to the predicate,
without checking its valid or not. Dereferencing a nullptr is undefinied behaviour.

For unknown reason no crash poped up in the unit tests.

This will fix #15, reported by sbalogh.
Adjusted imply of empty check, empty is, when begin() == end()