fixed registration of ctor which takes arguments by value
The problem was the implementation didn't worked for VS2013,
because of the lack of expression SFINAE support. Work better in VS2015.
However, VS 2013 should still be supported.

Replaced "is_callable" with a new trait called "is_functor".
Because there is already in C++17 an official trait with the name "is_callable".
This trait works also in VS2013.

This trait was provided by K-ballo.

The policy functions for creating "paramter_infos" and "default_arguments",
does now not depend anymore on the "is_function" trait (which depended on "is_callable")
we forward the information now from the caller (bind class) to these functions.
Explicit better then implicit.

Unit test added.