added 'create_invalid_item' function template
Now the creation of an invalid item goes with this method,
instead of calling 'create_item' with a nullptr.
This is more explicit and very important, we can move the creation of the
f the static iinvalid base wrapper item inside this function.
So we can make sure the static object is really created and ready.