Added completly new approach for storing the wrapper item
All data allocated with new in a module (DLL, so, ...)
should not be moved to the main RTTR library.
Instead the RTTR library will now store only the pointers to the data (register).
The data is now saved in a singleton (registration_manager) which is created per module (static global in function template)
When this manager object goes out of scope, the data will be unregistered before destroying.
This approach has the advantage, that the 'new' and 'delete' of the std::unique_ptr<T>
from the same STL library implementation will be used.
Instead of mixing it (this happens when I move the unique_ptr across module boundaries).