fixed runtime error on MacOSX, because of load/unload of libraries
The problem was that the invalid type_data object was created inside the library module
statically. Because we switched to using the pointer compare of types, instead of using
and separate ID, this lead to problems when comparing two invalid types.
They were the same.

Which specific code which lead to the problem is unclear. However, this behaviour was wrong
and is now fixed.

Additional: no clang memory or address sanitizer could find the source.
The problem was only poping up in MacOSX with apple clang.