fixed invalid memory access under macOSx with clang
Actually the problem was also present under linux, (with gcc or clang).
This situation was following:
1. A plugin was loaded with type "int[100]" and registered
2. The plugin was unloaded => "int[100]" was removed from the type system and deleted
3. The type[100] was used in the unit_test later and was accessing the memory
of the already deleted type_data

The reason for this behavioue lies in the usage of local static member variables
(which is used to cache the retrieving and registering of type_data information).
This variable was process wide unique, this lead to problems when unloading a plugin.
The static was no reinitialized when unloading the plugin.
So the variable was pointing to type_data information which was alreadyl deleted.

Now we are using RTTR_LOCAL to hide the symbols for exporting. This fixes the issues.