fixed not working invoke of global properties via visitor
The problem, was the global property class template was instantiated previously (type_prop_meth_invoke.cpp)
without a visitor. Then later where a visitor was available, the new instantiation was discarded by the linker (ODR).

In order to fix this, the visitor_list has to be provided during the class instantiation (so during registration process)
This commit will fix this.